requirements.txt
# Core
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0

# Database (SQLite per iniziare)
sqlite3

# Testing
pytest==7.4.3
pytest-asyncio==0.21.1

# Utils
pydantic==2.5.0
requests==2.31.0
python-multipart==0.0.6
